Download and add a new, refreshed, and tidy metric.
Downloaded opioid involved fatal overdoses by year and resident
municipality. This is a different metric and refreshes the view to 2020.
It includes history from 2009 to 2020. This would enable tightening up
what `<5` means by making an assumption that rates can't change too
rapidly: some of these are more likely to be 1s and others more likely
to be 4s. Whether the symbol is intermingled with 0s (Barrington) or
numbers >= 5 (Coventry) provides information on that, which would
improve the intervals and sorts compared to imputing a 2.5.

Source:

https://ridoh-overdose-surveillance-rihealth.hub.arcgis.com/datasets/municipal-count-of-opioid-involved-fatal-overdose-by-year-resident-municipality/explore